06-advanced-integrations-and-demos/fs-smartsheet/pyproject.toml
[project]
name = "fs-smartsheet"
version = "1.0.0"
description = "Abstraction layer between Finite State platform and Smartsheet"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "smartsheet-python-sdk>=3.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "filelock>=3.0",
]

[project.scripts]
fs-smartsheet = "fs_smartsheet.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fs_smartsheet"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "mypy>=1.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
# warnings.filterwarnings() must run before any smartsheet import
"src/fs_smartsheet/cli.py" = ["E402"]
"src/fs_smartsheet/__init__.py" = ["E402"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

# Relax strict checks that the existing codebase doesn't satisfy yet.
# These can be tightened incrementally.
disallow_untyped_defs = false
disallow_untyped_calls = false
warn_return_any = false

[[tool.mypy.overrides]]
module = "fs_smartsheet.cli"
# CLI module uses Typer decorators that are hard to fully annotate
disallow_untyped_decorators = false
